Refactor combat result enums and structures for clarity and consistency
- Updated `AttackResult` enum to include more descriptive variants: `NotAvailable`, `Success`, and `Fail`.
- Enhanced documentation for `AffectResult` to clarify possible outcomes and effects.
- Added detailed descriptions for various effect types in `EffectResult`, including permanent and temporary modifications.
- Introduced `RoundEffectResult` struct to encapsulate the source, target, and effect of round-based actions.
- Improved documentation across multiple structs and enums in the combat module for better understanding.

Enhance arena credits management interface

- Expanded `IArenaCredits` trait with detailed documentation for energy and credit management functions.
- Added utility functions for adding credits and consuming energy/credits with comprehensive argument descriptions.
- Introduced `ArenaCredit` and `Energy` structs to encapsulate user credit and energy data.

Revise attack and effect structures in loadout

- Updated `Attack` struct to clarify fields and improve documentation.
- Enhanced `Effect` struct with detailed descriptions of target, duration, and effect types.
- Introduced new enums for damage types and effect durations to improve clarity.

Refactor attributes and modifiers for combatants

- Added detailed documentation for ability scores, modifiers, resistances, and vulnerabilities.
- Introduced new structs for partial attributes and internal calculations to enhance attribute management.

Update loadout structures and scripts

- Enhanced `LoadoutInput` and `AttackSlot` structs with comprehensive field descriptions.
- Improved JavaScript scripts for parsing attributes and handling errors during execution.

+++ packages/arcade/src/attempt.cairo
@@ -15,6 +15,11 @@ pub type CombatNodePath = StoragePath<Mutable<CombatNode>>;
 pub type AttemptNodePath = StoragePath<Mutable<AttemptNode>>;
 
 
+/// Represents the current phase of an arcade attempt.
+//// * `None` - The attempt has not started.
+/// * `Active` - The attempt is currently active.
+/// * `PlayerWon` - The player has won the attempt.
+/// * `PlayerLost` - The player has lost the attempt.
 #[derive(Drop, Copy, Introspect, PartialEq, Serde, starknet::Store, Default)]
 pub enum ArcadeProgress {
     #[default]
@@ -38,12 +43,22 @@ impl CombatProgressIntoArcadePhase of Into<CombatProgress, ArcadeProgress> {
 }
 
 
+/// Represents an arcade attempt by a player.
+//// # Fields
+/// * `player` - The address of the player making the attempt.
+/// * `abilities` - The abilities of the player during the attempt.
+/// * `token_hash` - A hash representing the player's token.
+/// * `health_regen` - The health regeneration amount.
+/// * `expiry` - The timestamp when the attempt expires.
+/// * `phase` - The current phase of the arcade attempt.
+/// * `stage` - The current stage of the attempt.
+/// * `respawns` - The number of respawns the player has used.
 #[derive(Drop, Serde)]
 pub struct Attempt {
     pub player: ContractAddress,
     pub abilities: Abilities,
     pub token_hash: felt252,
-    pub health_regen: u32,
+    pub health_regen: u8,
     pub expiry: u64,
     pub phase: ArcadeProgress,
     pub stage: u32,
